#24 [refactor] viewModel 수정
